Move ECS plugin factory functions to plugin module
We'd like to keep the plugins as self-contained as possible so put
the functions to create the plugin and its associated object-graph
into the module containing the logic.

JIRA: PLAT-834
